apps.py for plugin being properly discovered
